--- analysis/tesselation.py
@@ -1,4 +1,0 @@
-from tesspy import Tessellation
-
-ffm = Tessellation("Dresden")
-
